% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzacaratteri.R
\name{normalizzacaratteri}
\alias{normalizzacaratteri}
\title{normalizzacaratteri}
\usage{
normalizzacaratteri(testo, fixed = TRUE)
}
\arguments{
\item{testo}{Un set di testi da cui rimuovere caratteri escape e punteggiatura.}

\item{fixed}{Logical. Se \code{TRUE}, il pattern e' una stringa da matchare com'e'. 
Esclude tutti gli argomenti in conflitto.}
}
\value{
Un set di testi processati.
}
\description{
\code{normalizzacaratteri} Sostituisce caratteri escape e punteggiatura in \code{testo} con uno spazio vuoto.
}
\examples{

 testo<-c("\t","\r") 
 normalizzacaratteri(testo)
 
 
}
\author{
Livio Finos
}
