% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzahtml.R
\name{normalizzahtml}
\alias{normalizzahtml}
\title{normalizzahtml}
\usage{
normalizzahtml(testo, perl = TRUE, fixed = TRUE)
}
\arguments{
\item{testo}{Un set di testi da preprocessare.}

\item{perl}{Logical. Se \code{TRUE}, utilizza regex compatibili con Perl-compatible.}

\item{fixed}{Logical. Se \code{TRUE}, il pattern e' una strings da matchare com'e'. 
Esclude tutti gli argomenti in conflitto.}
}
\value{
Un set di testi in cui gli URL sono stati sostituiti con la parola chiave WWWURLWWW.
}
\description{
\code{normalizzahtml} Sostituisce link e URL in \code{testo} con la parola chiave WWWURLWWW.
}
\examples{

 testo<-c("http:textwiller.com","www.textwiller.com") 
 normalizzahtml(testo)
 
 
}
\author{
Livio Finos
}
