% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzapunteggiatura.R
\name{normalizzapunteggiatura}
\alias{normalizzapunteggiatura}
\title{normalizzapunteggiatura}
\usage{
normalizzapunteggiatura(
  testo,
  removeUnderscore = TRUE,
  perl = TRUE,
  fixed = TRUE
)
}
\arguments{
\item{testo}{Un set di testi da cui rimuovere caratteri escape e punteggiatura.}

\item{removeUnderscore}{Logical. Se \code{TRUE}, rimuove gli undescore "_" dai testi.}

\item{perl}{Logical. Se \code{TRUE}, utilizza regex compatibili con Perl.}

\item{fixed}{Logical. Se \code{TRUE}, il pattern e' una stringa da essere matchata com'e'. 
Esclude tutti gli argomenti in conflitto.}
}
\value{
Un set di testi processati.
}
\description{
\code{normalizzapunteggiatura} Rimuove caratteri escape e punteggiatura da un set di testi.
}
\examples{

 testo<-c(testo<-c("@ ","@retweet","# ","#ciao")) 
 normalizzapunteggiatura(testo)
 

}
\author{
Livio Finos
}
