% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzaslang.R
\name{normalizzaslang}
\alias{normalizzaslang}
\title{normalizzaslang}
\arguments{
\item{testo}{Oggetto \code{character vector} di testi da normalizzare}

\item{perl}{Logico. Se \code{TRUE}, vengono utilizzate regex compatibili con Perl.}
}
\value{
Un oggetto \code{character vector} di testi in cui le espressioni 
gergali italiane sono state sostituite da una parola chiave.
}
\description{
Sostituisce esperessioni gergali italiane in \code{testo}
con una parola chiave corrispondente.
}
\details{
La parola chiave pu昼㸲 essere la parola stessa normalizzata (es., nn = non)
oppure un'emoticon (es., zzz = EMOTEZZZ).
}
\examples{
testo <- c("grandissima", "aaa", "nun")
normalizzaslang(testo)

}
\author{
Livio Finos
}
