% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzaTesti.R
\name{normalizzaTesti}
\alias{normalizzaTesti}
\alias{normalizzacaratteri}
\alias{normalizzaemoticon}
\alias{normalizzahtml}
\alias{normalizzapunteggiatura}
\alias{tryTolower}
\alias{stopwords_ita}
\alias{removeStopwords}
\alias{preprocessingEncoding}
\title{Funzioni di normalizzazione del testo}
\usage{
normalizzaTesti(
  testo,
  tolower = TRUE,
  normalizzahtml = TRUE,
  normalizzacaratteri = TRUE,
  normalizza_emoticon = TRUE,
  normalizzapunteggiatura = TRUE,
  normalizzaslang = TRUE,
  fixed = TRUE,
  perl = TRUE,
  preprocessingEncoding = TRUE,
  encoding = "UTF-8",
  sub = "",
  contaStringhe = c("\\\\?", "\\\\!", "@", "#", "(\\200|euro)", "(\\\\$|dollar)",
    "SUPPRESSEDTEXT"),
  suppressInvalidTexts = TRUE,
  verbatim = TRUE,
  remove = TRUE,
  removeUnderscore = FALSE
)
}
\arguments{
\item{testo}{Vettore contentente il testo.}

\item{tolower}{\code{TRUE} di default.}

\item{normalizzahtml}{\code{TRUE} di default.}

\item{normalizzacaratteri}{\code{TRUE} di default.}

\item{normalizza_emoticon}{\code{TRUE} di default.}

\item{normalizzapunteggiatura}{\code{TRUE} di default.}

\item{normalizzaslang}{\code{TRUE} di default.}

\item{fixed}{vedi \code{\link[base:gsub]{base:gsub}}. Preferibilmente non
usare l'opzione.}

\item{perl}{vedi \code{\link[base:gsub]{base:gsub}}. Preferibilmente non
usare l'opzione.}

\item{preprocessingEncoding}{Logical.}

\item{encoding}{\code{"UTF-8"} default. Se \code{FALSE} evita la
conversione.}

\item{sub}{Stringa character. Se non NA, utilizzata per sostituire byte non convertibili nell'input.
Vedi anche il parametro \code{sub} in \code{iconv}.}

\item{contaStringhe}{Stringhe da contare nei documenti. Default: \code{
c("\\?","\\!","#","@", "(€|euro)","(\\$|dollar)","SUPPRESSEDTEXT")}.}

\item{suppressInvalidTexts}{Sostituisce con \code{"SUPPRESSEDTEXT"} le
stringhe con mutibyte non validi (che produrrebbero verosimilmente errori
nelle successive normalizzazioni). Default \code{TRUE}.}

\item{verbatim}{Mostra statitiche durante il processo. \code{TRUE} di default.}

\item{remove}{\code{TRUE} by default. Un vettore di parole da rimuovere.}

\item{removeUnderscore}{Rimuovere gli underscore?}

\item{ifErrorReturnText}{Cosa ritornare per testi con encoding errato.}

\item{stopwords}{Lista di parole da escludere dall'analisi. \code{stopwords_ita} di default.}
}
\value{
Per \code{normalizzaTesti} l'output e' il vettore di testi
normalizzati. La tabella dei conteggi specificati in \code{contaStringhe}
e' assegnata come tabella \code{counts} tra gli \code{attributes} del
vettore stesso.

Per tutte le altre funzioni, l'output e' un \code{vector} della stessa
lunghezza di \code{testo} ma con testi normalizzati.
}
\description{
Normalizzazione di varie carratteristiche del testo in preparazione all'analisi.
}
\details{
\code{stopwords_ita} e' una lista di stopword italiane.
}
\note{
%% ~~further notes~~
}
\examples{
testoNorm <- normalizzaTesti(c('ciao bella!','www.associazionerospo.org','noooo, che grandeeeeee!!!!!','mitticooo', 'mai possibile?!?!'))
testoNorm
attr(testoNorm,"counts")

}
\author{
Dario Solari, Livio Finos, Maddalena Branca, Mattia Da Pont
}
\keyword{~kwd1}
\keyword{~kwd2}
