% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTHound.R
\name{RTHound}
\alias{RTHound}
\title{RTHound}
\usage{
RTHound(
  testo,
  S = 500,
  L = 100,
  hclust.dist = 100,
  hclust.method = "complete",
  showTopN = 5,
  dist = "levenshtein",
  verbatim = TRUE
)
}
\arguments{
\item{testo}{Tweet o vettore di testo.}

\item{S}{Numero di tweet (o testi) per ogni sottoinsieme. \code{500} di
deafault.}

\item{L}{Numero di tweet (o testi) appartenenti al precedente sottoinsieme
da inserire nell'analisi. \code{100} di default.}

\item{hclust.dist}{Scalare numerico contenente l'altezza a cui tagliare gli alberi.
\code{100} di deafault.}

\item{hclust.method}{Metodo di agglomerazione. Dovrebbe essere un'abbreviazione
non ambigua scelta tra \code{"ward"}, \code{"single"},
\code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} o
\code{"centroid"}. \code{"complete"} di default.}

\item{showTopN}{Numero di tweet piu' frequenti da mostrare. \code{5} di
deafault.}

\item{dist}{\code{"levenshtein"} e' il default. \code{"profile"} e' l'altro 
(piu' rapido) valore accettato.}

\item{verbatim}{Logical.}
}
\value{
\code{RTHound} sostituisce i tweet appartenenti allo stesso cluster
con quello piu' vecchio, identificandoli come retweet, e restituisce una lista
dei retweet piu' frequenti (\code{top}).
}
\description{
Identifica i retweet piu' frequenti tramite clustering gerarchico applicato 
ad una matrice delle distanze di Levenshtein.
}
\details{
\code{RTHound} divide il testo in \code{testo} in sottoinsiemi di lunghezza \code{S} 
(dal secondo sottoinsieme incorpora anche \code{L} tweet del sottoinsieme precedente).
La funzione calcola una matrice di dissomigliana basata sulla distanza di Levenshtein per
ogni sottoinsieme e raggruppa i tweet tramite un algoritmo di clustering gerarchico.
}
\note{
%% ~~further notes~~
}
\examples{

 \dontrun{
 testo=c(
 "RT @LAVonlus: Tre miti da sfatare sulla #vivisezione. Le risposte  ai luoghi comuni della sperimentazione animale  http://t.co/zHSfam16DT",
 "Tre miti da sfatare sulla #vivisezione. Le risposte  ai luoghi comuni della sperimentazione animale  http://t.co/zHSfam16DT",
 "RT @LAVonlus: Tre miti da sfatare sulla #vivisezione. Le risposte  ai luoghi comuni della sperimentazione animale  http://t.co/zHSfam16DT",
 "RT @orianoPER: La #sperimentazioneanimale è inutile perché non predittiva per la specie umana. MEDICI ANTI #VIVISEZIONE- LIMAV http://t.co/" ,
 "La #sperimentazioneanimale è inutile perché non predittiva per la specie umana. MEDICI ANTI #VIVISEZIONE- LIMAV http://t.co/3MwubXIH8g",
 "RT @orianoPER: La #ricerca in #Medicina con #sperimentazioneanimale non e' predittiva per la specie umana. MEDICI ANTI #VIVISEZIONE http://t",
 "RT @HuffPostItalia: Il Governo italiano non fermi la sperimentazione animale. Intervista a Elena Cattaneo http://t.co/q1dm430a9j",
 "RT @HuffPostItalia: \"Il Governo italiano non fermi la sperimentazione animale\". Intervista a Elena Cattaneo http://t.co/q1dm430a9j",
 "\"Il Governo italiano non fermi la sperimentazione animale\". Intervista a Elena Cattaneo http://t.co/q1dm430a9j",
 "RT @orianoPER: @EnricoLetta LA #VIVISEZIONE NON SERVE: PAROLA DI GLAXO-APTUIT http://t.co/mtsHJjDIvu #StopVivisection #SperimentazioneAnima&")
 
 testo=RTHound(testo, S = 3, L = 1, 
                 hclust.dist = 100, hclust.method = "complete",
                 showTopN=3)

}

}
\seealso{
\code{hclust}
}
\author{
Federico Ferraccioli, Livio Finos
}
\keyword{~kwd1}
\keyword{~kwd2}
