% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment}
\alias{sentiment}
\alias{sentimentVocabularies}
\alias{vocabolariMadda}
\alias{vocabolarioMattivio}
\title{Sentiment analysis}
\usage{
sentiment(
  text,
  algorithm = "Mattivio",
  vocabularies = NULL,
  normalizzaTesti = TRUE,
  get_labels = TRUE,
  ...
)
}
\arguments{
\item{text}{Vettore di testi.}

\item{algorithm}{\code{"Mattivio"} (default), \code{"Maddalena"}, o una funzione
che restituisce il punteggio.}

\item{vocabularies}{\code{vocabolarioMattivio} di default se \code{algorithm == "Mattivio"}; 
\code{vocabolariMadda} di default se \code{algorithm == "Maddalena"}, oppure un 
oggetto usato dall'algoritmo.}

\item{normalizzaTesti}{\code{TRUE} di default.}

\item{get_labels}{\code{TRUE} di default. Se \code{FALSE}, assegna un punteggio
quantitativo; Se \code{TRUE}, assegna le etichette -1, 0, +1 (i.e., \code{sign(score)})}
}
\value{
Un array contenente un singolo valore numerico per ogni elemento del vettore \code{text}.
}
\description{
Assegna un punteggio sentiment (sentiment score) per ogni testo in \code{text}.
}
\note{
%% ~~further notes~~
}
\examples{

sentiment(c("ciao bella", "mi piaci", "wow!!","good","casa", "farabutto!","ti odio"))

}
\author{
Maddalena Branca, Mattia Da Pont, Livio Finos
}
\keyword{~kwd1}
\keyword{~kwd2}
